+++ bacnet-stack/src/mstp.c
@@ -714,7 +714,7 @@ bool MSTP_Master_Node_FSM(
                 uint8_t frame_type = mstp_port->OutputBuffer[2];
                 uint8_t destination = mstp_port->OutputBuffer[3];
                 RS485_Send_Frame(mstp_port,
-                    (uint8_t *) & mstp_port->OutputBuffer[0], length);
+                    (uint8_t *) & mstp_port->OutputBuffer[0], (uint16_t)length);
                 mstp_port->FrameCount++;
                 switch (frame_type) {
                     case FRAME_TYPE_BACNET_DATA_EXPECTING_REPLY:
@@ -1049,7 +1049,7 @@ bool MSTP_Master_Node_FSM(
                 /* then call MSTP_Create_And_Send_Frame to transmit the reply frame  */
                 /* and enter the IDLE state to wait for the next frame. */
                 RS485_Send_Frame(mstp_port,
-                    (uint8_t *) & mstp_port->OutputBuffer[0], length);
+                    (uint8_t *) & mstp_port->OutputBuffer[0], (uint16_t)length);
                 mstp_port->master_state = MSTP_MASTER_STATE_IDLE;
             } else {
                 /* DeferredReply */
